+++ packages/cardano-services-client/src/WebSocket.ts
@@ -1,18 +1,16 @@
 /* eslint-disable unicorn/prefer-add-event-listener */
 
 import {
-  AsyncReturnType,
   Cardano,
   EpochInfo,
   EraSummary,
-  NetworkInfoMethods,
+  HealthCheckResponse,
   NetworkInfoProvider,
+  Provider,
   ProviderError,
   ProviderFailure,
   StakeSummary,
   SupplySummary,
-  WSMessage,
-  WsProvider,
   createSlotEpochInfoCalc
 } from '@cardano-sdk/core';
 import { Logger } from 'ts-log';
@@ -21,6 +19,19 @@ import { fromSerializableObject } from '@cardano-sdk/util';
 import WebSocket from 'isomorphic-ws';
 
 const NOT_CONNECTED_ID = 'not-connected';
+
+export type AsyncReturnType<F extends () => unknown> = F extends () => Promise<infer R> ? R : never;
+
+export type NetworkInfoMethods = Exclude<keyof NetworkInfoProvider, 'healthCheck'>;
+export type NetworkInfoResponses = { [m in NetworkInfoMethods]: AsyncReturnType<NetworkInfoProvider[m]> };
+
+export interface WSMessage {
+  /** The client id assigned by the server. */
+  clientId?: string;
+
+  /** Latest value(s) for the `NetworkInfoProvider` methods.*/
+  networkInfo?: Partial<NetworkInfoResponses>;
+}
 
 type WSStatus = 'connecting' | 'connected' | 'idle' | 'stop';
 
@@ -53,6 +64,35 @@ interface EpochRollover {
 const isEventError = (error: unknown): error is { error: Error } =>
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   typeof error === 'object' && !!error && (error as any).error instanceof Error;
+
+export class WsProvider implements Provider {
+  /** Emits the health state. */
+  public health$: Observable<HealthCheckResponse>;
+
+  private healthSubject$: ReplaySubject<HealthCheckResponse>;
+  private reason?: string;
+
+  constructor() {
+    this.health$ = this.healthSubject$ = new ReplaySubject<HealthCheckResponse>(1);
+    this.healthSubject$.next({ ok: false, reason: 'starting' });
+  }
+
+  protected emitHealth(reason?: string, overwrite?: boolean) {
+    if (!reason) {
+      this.reason = undefined;
+
+      return this.healthSubject$.next({ ok: true });
+    }
+
+    if (overwrite || !this.reason) this.reason = reason;
+
+    this.healthSubject$.next({ ok: false, reason: this.reason });
+  }
+
+  public healthCheck() {
+    return firstValueFrom(this.health$);
+  }
+}
 
 export class CardanoWsClient extends WsProvider {
   /** The client id, assigned by the server. */
