+++ packages/core/src/index.ts
@@ -5,4 +5,3 @@ export * from './Provider';
 export * from './util';
 export * from './errors';
 export * from './CardanoNode';
-export * from './WebSocket';
